Restructure test files
